[Feat] #641 - Post 공통 셀 UI 및 애니메이션 구현
